New files: test_arithmetic.py and test_doc.py
